// web/platform-admin-web/src/api/model/inviteResponseDtoInviteStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App API
 * The AI Base App Gateway API documentation
 * OpenAPI spec version: 1.0
 */

/**
 * 초대 상태
 */
export type InviteResponseDtoInviteStatus = typeof InviteResponseDtoInviteStatus[keyof typeof InviteResponseDtoInviteStatus];


export const InviteResponseDtoInviteStatus = {
  PENDING: 'PENDING',
  CANCELED: 'CANCELED',
  ACCEPTED: 'ACCEPTED',
  REJECTED: 'REJECTED',
} as const;