// web/platform-admin-web/src/api/model/memberResponseDtoRole.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App API
 * The AI Base App Gateway API documentation
 * OpenAPI spec version: 1.0
 */

/**
 * 권한
 */
export type MemberResponseDtoRole = typeof MemberResponseDtoRole[keyof typeof MemberResponseDtoRole];


export const MemberResponseDtoRole = {
  OWNER: 'OWNER',
  MANAGER: 'MANAGER',
  VIEWER: 'VIEWER',
} as const;